+++ History.md
@@ -9,10 +9,10 @@
   information about how dynamic `import(...)` works in Meteor, and how to
   use it in your applications.
 
-* The `meteor-babel` npm package has been upgraded to version 0.20.2, and
-  the `reify` npm package has been upgraded to version 0.8.3, fixing
+* The `meteor-babel` npm package has been upgraded to version 0.20.3, and
+  the `reify` npm package has been upgraded to version 0.9.2, fixing
   [issue #8595](https://github.com/meteor/meteor/issues/8595) and
-  improving compilation performance.
+  improving compilation and runtime performance.
 
 ## v1.4.4.2, 2017-05-02